requirements.txt
typer[all]>=0.9.0
azure-identity>=1.14.0
requests>=2.31.0
pyyaml>=6.0
rich>=13.0.0
azure-cli-core>=2.50.0
